push_fifth

+++ src/main/java/com/sast/atSast/mapper/ContestMapper.java
@@ -19,7 +19,8 @@ public interface ContestMapper {
 
     void createContest(Contest contest);
 
-    void updatepushLink(int contestId, String pushLink);
+    void updatepushLink(long contestId, String pushLink);
 
     List<TeamMember> getTeamById(long contestId);
+
 }
